src/requirements.txt
opencv-python
Pillow
numpy
keras
tensorflow
